requirements.txt
defusedxml==0.4.1
dj-database-url==0.4.1
django==1.10
django-allauth==0.27.0
django-braces==1.10.0
django-floppyforms==1.7.0
django-invitation==1.0
django-invitations==1.8
django-modalview==0.1.5
django-registration==2.1.2
django-registration-defaults==0.4.4
djangorestframework==3.4.7
gunicorn==19.6.0
oauthlib==2.0.0
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
South==1.0.2
tweepy==3.5.0
virtualenv==15.0.3
whitenoise==3.2.2
